pyproject.toml
[tool.poetry]
name = "study-buddy-llm"
version = "0.1.0"
description = ""
authors = ["Bruna Calistrate <bruna.calistrate@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
streamlit = "^1.35.0"
python-dotenv = "^1.0.1"
apify-client = "^1.7.0"
langchain = "0.1.17"
langchain-google-genai = "1.0.3"
langchain-community = "0.0.37"
langchain-core = "0.1.52"
langchain-pinecone = "0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.6"
taskipy = "^1.12.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.ruff]

select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

line-length = 80
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.isort]
profile = "black"
line_length = 80
extend_skip = ['migrations']

[tool.taskipy.tasks]
lint = 'ruff . && black --check . --diff'
format = 'black .  && isort .'
run = 'streamlit run app/Home.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'